SEACAS: Fix compiler warnings
(From Sierra)

Build/Test Cases Summary
Enabled Packages: SEACASAprepro, SEACASConjoin, SEACASEpu, SEACASExodiff, SEACASExomatlab, SEACASAprepro_lib, SEACASExodus, SEACASIoss
Disabled Packages: Panzer,PyTrilinos,SEACASExo2mat,SEACASMat2exo
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=120,notpassed=0 (30.31 min)
3) SERIAL_RELEASE_SS => passed: passed=122,notpassed=0 (26.74 min)
Other local commits for this build/test group: 1c2e5f7, 6a41c31, f3fb082, 3c183c1, a517ace, e49366a, 8ccf3d7, beae79e, f9a18d9

+++ packages/seacas/libraries/aprepro_lib/apr_builtin.cc
@@ -1218,7 +1218,7 @@ array *do_csv_array(const char *filename)
   while (std::getline(*file, line)) {
     std::vector<std::string> tokens;
     tokenize(line, delim, tokens);
-    for (size_t i=0; i < array_data->cols; i++) {
+    for (size_t i=0; i < (size_t)array_data->cols; i++) {
       if (i < tokens.size()) {
 	array_data->data[idx++] = atof(tokens[i].c_str());
       }
@@ -1228,7 +1228,7 @@ array *do_csv_array(const char *filename)
     }
     rows++;
   }
-  assert(rows == array_data->rows);
+  assert((int)rows == array_data->rows);
   return array_data;
 }
